Instance: CMSG_SET_DIFFICULTY_ID carries a uint16, not a uint32
This packet class was added on this branch (d363a89ba0, "Instance: Add realm-wide
encounter start/end, release control, difficulty and abandon vote packets") and the
reader has been wrong since. The 68275 client serializer at RVA 0x6A7840 does
`movzx edx, word ptr [msg+0x20]; call write_uint16 (0x33CCDC0)` and returns - the entire
body is two bytes. Reading a uint32 off it threw ByteBufferPositionException out of
Read() before HandleSetDifficultyID was ever called, so changing difficulty through this
opcode has never once worked. The sibling CMSG_SET_DUNGEON_DIFFICULTY (client RVA
0x5D7650) writes the same uint16 and MiscPackets already reads int16; this one was the
outlier.

The handler needs no change - it only feeds DifficultyID to sDifficultyStore.LookupEntry.

Transplanted from feature/gap-closers d81cd273b9, which fixed this on the integration
line only. This branch still carried the broken version, so a later merge could have
reinstated it.

No SQL. Nothing deployed.

+++ src/server/game/Server/Packets/InstancePackets.h
@@ -689,7 +689,12 @@ namespace WorldPackets
 
             void Read() override;
 
-            uint32 DifficultyID = 0;
+            // The 68275 client writes this as a uint16, not a uint32: the serializer at client RVA 0x6A7840
+            // does `movzx edx, word ptr [msg+0x20]; call write_uint16 (0x33CCDC0)` and nothing else, so the
+            // whole body is two bytes. Reading a uint32 off a two-byte packet threw ByteBufferPositionException
+            // and the handler never ran. The sibling CMSG_SET_DUNGEON_DIFFICULTY (client RVA 0x5D7650) uses the
+            // same uint16 write and MiscPackets already reads it as int16 - this one was the odd man out.
+            uint16 DifficultyID = 0;
         };
 
         class ToggleDifficulty final : public ClientPacket
